refactor(step): 공통 액션 step으로 분리
