+++ src/test/java/nextstep/subway/auth/acceptance/AuthAcceptanceStepTest.java
@@ -1,0 +1,21 @@
+package nextstep.subway.auth.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.auth.dto.TokenRequest;
+import org.springframework.http.MediaType;
+
+public class AuthAcceptanceStepTest {
+
+    public static ExtractableResponse<Response> 로그인_요청(TokenRequest tokenRequest) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(tokenRequest)
+                .when().post("/login/token")
+                .then().log().all()
+                .extract();
+    }
+
+}
