+++ src/test/java/nextstep/subway/path/PathAcceptanceStepTest.java
@@ -1,0 +1,28 @@
+package nextstep.subway.path;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class PathAcceptanceStepTest {
+    public static ExtractableResponse<Response> 최단_경로_조회_요청(String accessToken, Long sourceStationId, Long targetStationId) {
+        Map<String, Long> params = new HashMap<>();
+        params.put("source", sourceStationId);
+        params.put("target", targetStationId);
+
+        return RestAssured.given().log().all()
+                .auth().oauth2(accessToken)
+                .queryParams(params)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/paths")
+                .then().log().all()
+                .extract();
+    }
+
+
+
+}
